requirements.txt
python-binance
python-dotenv
requests
rich
loguru
aiohttp
